+++ limbus/app/patientconsentform/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 pcf = Blueprint("pcf", __name__)
 
-from . import views
+from . import routes
